backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-multipart
httpx>=0.24.0
pytest
sqlalchemy
psycopg2-binary
alembic
prometheus_client
rq
redis

passlib[bcrypt]
bcrypt>=4.0.0
python-jose[cryptography]
